Get Color Palette of the Image
Get Color Palette of the Image , only send the colors name and hex on the chat.

Color numbers in the palette is the default, 5.

Make Custom image for each color of the Palette

Make The Palatte Image and send

Code organization and last adjustments in the "Getcolor" command.

Some interactions and Help command

Initial comit of +more

+More Command End

Update Help msg